+++ source/Energinet.DataHub.MarketParticipant.Client/source/Energinet.DataHub.MarketParticipant.Client/TokenClient.cs
@@ -14,17 +14,25 @@
 
 using System.Threading.Tasks;
 using Energinet.DataHub.MarketParticipant.Client.Models;
+using Flurl.Http;
 
 namespace Energinet.DataHub.MarketParticipant.Client
 {
-    /// <summary>
-    /// BFF client for DataHub token retrieval
-    /// </summary>
-    public interface IMarketParticipantTokenClient
+    public sealed class TokenClient : ITokenClient
     {
-        /// <summary>
-        /// Retrieves a DataHub token
-        /// </summary>
-        Task<GetTokenResponseDto> GetTokenAsync(GetTokenRequestDto request);
+        private readonly IFlurlClient _httpClient;
+
+        public TokenClient(IFlurlClient client)
+        {
+            _httpClient = client;
+        }
+
+        public async Task<GetTokenResponseDto> GetTokenAsync(GetTokenRequestDto request)
+        {
+            var response = await ValidationExceptionHandler
+                .HandleAsync(
+                    () => _httpClient.Request("token").PostJsonAsync(request)).ConfigureAwait(false);
+            return await response.GetJsonAsync<GetTokenResponseDto>().ConfigureAwait(false);
+        }
     }
 }